Update LICENSE
This is an acknowledgement that my license choice was overzealous for just a bunch of scripts. Switching to BSD 3 clause, again all commits are also bound to GPLv3 and AGPLv3 and may be used under that license. As the sole license holder, I cannot revoke those old licenses but may provide these under a new one.